requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
Flask-Script
Flask-Login
psycopg2-binary
gunicorn
flask-debugtoolbar
flake8
alembic
Flask-Mail
python-dateutil
python-editor
Werkzeug
WTForms
